Refector the state
Now there are objects in the state, thanks to this we can control isHeld
value(if it is true, this number will not change in the next roll). Also
change the name of an array for more code clarity. Now as a number prop
el.number is passed - this change results from the changes in state
structure